Finished generating events

+++ src/main/java/JSONData/Location.java
@@ -1,19 +1,19 @@
 package JSONData;
 
 public class Location {
-    private double latitude, longitude;
+    private float latitude, longitude;
     private String city, country;
 
-    public Location(double latitude, double longitude, String country, String city) {
+    public Location(float latitude, float longitude, String country, String city) {
         this.latitude = latitude;
         this.longitude = longitude;
         this.country = country;
         this.city = city;
     }
 
-    public double getLatitude() {return latitude;}
+    public float getLatitude() {return latitude;}
 
-    public double getLongitude() {return longitude;}
+    public float getLongitude() {return longitude;}
 
     public String getCity() {return city;}
